+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Changelog
 Next
 ----
 
-- Add the capability of sending directories to ``Nodes`` via ``Node.send_file``.
+- Add the capability of sending a directory to a ``Node`` via ``Node.send_file``.
 - Add ``ip_detect_path`` to the each ``ClusterBackend`` as a property and to each install DC/OS function as a parameter.
 
 2018.07.23.1